display: take the fbdev dkmd and ukmd trees and fix the wiring
The DPU implementation lives in drivers/video/fbdev/{dkmd,ukmd} in the source
tree; the platform_source/display/drivers/video copy holds the same files with
identical Makefile hashes. Bring those directories over, stop descending into
platform_source/display/drivers/dpu_drm, whose child Makefiles reference
sources that do not exist (it only carries the Kconfig symbols), and source
the ported Kconfigs so DKMD_ENABLE and UKMD_ENABLE exist.

Source: HarmonyOS6_Hi36C0_Kernel/drivers/video/fbdev/{dkmd,ukmd,akmd}

+++ drivers/video/fbdev/dkmd/dpu/begonia/composer/online/ppc/dpu_comp_ppc.c
@@ -14,6 +14,7 @@
 #include "dpu_panel_partial_ctl.h"
 #include "dkmd_log.h"
 #include "dkmd_notify.h"
+#include "dksm_debug.h"
 
 void dpu_comp_ppc_ctrl_setup(struct dpu_composer *dpu_comp, struct comp_online_present *present)
 {
@@ -59,7 +60,7 @@ int32_t dpu_comp_preprocess_set_active_rect(struct dpu_composer *dpu_comp, uint3
 	pinfo->ppc_config_id_record = ppc_config_id;
 
 	event.data = NULL;
-	event.value = pinfo->ppc_config_id_record;
+	event.value = (int32_t)pinfo->ppc_config_id_record;
 	dkmd_notifier_call_chain(DKMD_EVENT_SET_ACTIVE_RECT, (void *)&event);
 
 	if (alsc_need_update)
